Time: 26 ms (55.73%) | Memory: 38.8 MB (19.49%) - LeetSync

+++ 485-max-consecutive-ones/max-consecutive-ones.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int findMaxConsecutiveOnes(vector<int>& nums) {
-        int count=0;
-        int maximum=0;
-        int n=nums.size();
+        long long count=0;
+        long long maximum=0;
+        long long n=nums.size();
         for(int i=0;i<n;i++){
             if(nums[i]==1){
                 count++;
